// frontend/src/types/api.ts
export interface components {
    schemas: {
        Problem: {
            /** Format: uri */
            type?: string;
            title?: string;
            status?: number;
            detail?: string;
            instance?: string;
        };
        User: {
            /** Format: uuid */
            userId?: string;
            /** Format: email */
            email?: string;
            displayName?: string;
            /** @enum {string} */
            role?: "user" | "moderator" | "admin";
        };
        /** @enum {string} */
        RoomVisibility: "public" | "private";
        Room: {
            /** Format: uuid */
            roomId?: string;
            name?: string;
            description?: string;
            visibility?: components["schemas"]["RoomVisibility"];
            isArchived?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            updatedAt?: string;
        };
        RoomMember: {
            /** Format: uuid */
            roomId?: string;
            /** Format: uuid */
            userId?: string;
            /** @enum {string} */
            roleInRoom?: "owner" | "admin" | "member";
            /** Format: date-time */
            joinedAt?: string;
        };
        Message: {
            /** @description ULID */
            messageId?: string;
            /** Format: uuid */
            roomId?: string;
            /** Format: uuid */
            userId?: string;
            body?: string;
            /** @enum {string} */
            status?: "active" | "deleted";
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            editedAt?: string | null;
        };
        MessagePost: {
            /** Format: uuid */
            roomId: string;
            body: string;
            /** @description ULID generated on client */
            clientMsgId?: string;
        };
        MessagePatch: {
            body?: string;
        };
        InviteToken: {
            token?: string;
            /** Format: uuid */
            roomId?: string;
            /** Format: date-time */
            expiresAt?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}